[JSC] Fix shiftI64 spilling only on one side of a branch on 32-bit https://bugs.webkit.org/show_bug.cgi?id=298046
Reviewed by Justin Michaud.

The program import-mutable-global.js generates the following in armv7:

1. if shift is zero jump to end
2. if shift is less than 32 do:
2.1 spill r0, r1 to use as scratch
2.2 do the shift
...
4. end

however, if we take shift is zero and we jump to end, BBQ still thinks we did
the spill and reloads r0, r1, loading garbage.

This PR fixes the issue by doing the spill as soon as we enter the function, so
we always reload the right values.

* Source/JavaScriptCore/wasm/WasmBBQJIT32_64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::shiftI64Helper):

Canonical link: https://commits.webkit.org/299328@main