[css-anchor-position-1] Last-successful position fallback doesn't apply on anchoreum.com rdar://158365856 https://bugs.webkit.org/show_bug.cgi?id=297438
Reviewed by Antti Koivisto.

If a PositionTryFallback is a position-area option, the position-area is
stored in a RefPtr<const StyleProperties>. The current operator== for
PositionTryFallback is default, so it'd only check whether the StyleProperties
pointers are the same. Fix it to look inside the StyleProperties and check
if the CSSValue of position-area are equal.

Also improve operator<< of PositionTryFallback to print out more debug
information.

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-fallbacks-position-area-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-change-fallbacks-position-area.html: Added.
* Source/WebCore/rendering/style/PositionTryFallback.cpp:
(WebCore::Style::operator==):
(WebCore::Style::operator<<):
* Source/WebCore/rendering/style/PositionTryFallback.h:
* Source/WebCore/style/StyleBuilderConverter.h:
(WebCore::Style::BuilderConverter::convertPositionTryFallbacks):
- Use ImmutableStyleProperties::createStyleDeduplicating to deduplicate the
  returning ImmutableStyleProperties, if an identical one was created before.

Canonical link: https://commits.webkit.org/299367@main