Introduce AudioVideoRenderer class. https://bugs.webkit.org/show_bug.cgi?id=297961 rdar://159279717
Reviewed by Eric Carlson.

And make MediaPlayerPrivateWebM use its AVFoundation implementation
AudioVideoRendererAVFObjC.

AudioVideoRenderer abstract the entire interaction with the system framework
required to decode and render a video or MediaSample and handle A/V synchronization.

By using an AudioVideoRenderer a MediaPlayerPrivate is only left to handle
the demuxing of the media file.

By having this new layer separation, we will be able in a future change
to have the MediaPlayerPrivate runs with its demuxer in the Content Process
while the AudioVideoRenderer runs remotely.

For now, there's no visible functional changes except on visionOS.
This current AudioVideoRendererAVFObjC is only usable on systems where
WebCoreDecompressionSession is preferred to decode MediaSample.

On visionOS the transition between an AVSampleBufferDisplayLayer and
a AVSampleBufferVideoRenderer is fully handled by AudioVideoRenderer rather
than the MediaPlayerPrivate. It no longer requires to perform internal
seeks and re-enqueuing samples.
The same improvement also occurs when the video element will be added/removed
from the DOM, the transition is instantaneous without having to perform
similar sets of operations as we no longer need to switch between an AVSBDL
and a WebCoreDecompressionSession.

Covered by existing tests.

Fullscreen, PiP and docking transitions manually tested on iPad and Vision Pro
as we don't have automated coverage for those.

* Source/WTF/wtf/LoggerHelper.h: Add DEBUG_LOG_WITH_THIS and DEBUG_LOG_WITH_THIS_IF_POSSIBLE macro.

Canonical link: https://commits.webkit.org/299373@main