[Tools][WPE] Plug run-mvt-tests script into webkitpy display-server https://bugs.webkit.org/show_bug.cgi?id=298108
Reviewed by Nikolas Zimmermann.

This plugs the script run-mvt-tests into the webkitpy display-server facility.
That way it becomes easier to run the tests headless in GTK (via the xvfb
display-server) and allows also to run the tests with WPE on any of the wayland
or headless display-servers.

* Tools/Scripts/run-mvt-tests:
(parse_args):
* Tools/Scripts/webkitpy/webdriver_tests/webdriver_test_runner.py:
(WebDriverTestRunner.__init__):

Canonical link: https://commits.webkit.org/299387@main