REGRESSION(299129@main): [ iOS Debug ]ASSERTION FAILED: !isInAuxiliaryProcess() || MediaSessionHelper::sharedHelper().presentedApplicationPID() https://bugs.webkit.org/show_bug.cgi?id=298131 rdar://159478772
Reviewed by Youenn Fablet.

The assertion added to AudioSessionCocoa::tryToSetActiveInternal in 299129@main is invalid on
platforms where media capability grants are enabled, since in that case it is expected for
presentedApplicationPID to not be set. Since it's not possible to check the value of a web
preference in AudioSession, restricted the assertion at compile time to platforms where
ENABLE(EXTENSION_CAPABILITIES) is false.

However, this assertion did expose a bug in 299129@main where
UserMediaCaptureManagerProxySourceProxy would activate the AudioSession without first setting a
presentedApplicationPID (on platforms where that's necessary). Resolved that by setting a Function
on UserMediaCaptureManagerProxySourceProxy that is called prior to activating the audio session.
UserMediaCaptureManagerProxy sets the function to a lamda that calls
GPUConnectionToWebProcess::providePresentingApplicationPID with the proxy's pageIdentifier.

* LayoutTests/platform/ios/TestExpectations:
* Source/WebCore/platform/audio/cocoa/AudioSessionCocoa.mm:
(WebCore::AudioSessionCocoa::tryToSetActiveInternal):
* Source/WebKit/GPUProcess/GPUConnectionToWebProcess.cpp:
(WebKit::GPUConnectionToWebProcess::providePresentingApplicationPID const):
(WebKit::providePresentingApplicationPID): Deleted.
* Source/WebKit/GPUProcess/GPUConnectionToWebProcess.h:
* Source/WebKit/GPUProcess/webrtc/UserMediaCaptureManagerProxy.cpp:
(WebKit::UserMediaCaptureManagerProxy::createMediaSourceForCaptureDeviceWithConstraints):
(WebKit::UserMediaCaptureManagerProxy::clone):
* Source/WebKit/GPUProcess/webrtc/UserMediaCaptureManagerProxy.h:

Canonical link: https://commits.webkit.org/299430@main