Schedule task on buffered PerformanceObserver https://bugs.webkit.org/show_bug.cgi?id=297765 rdar://158950657
Reviewed by Chris Dumez and Ryan Reno.

PerformanceObserver::observe() with `buffered: true` now schedules a task to
run all callbacks instead of invoking the relevant callbacks directly. The
new behavior is more conformant; see
https://www.w3.org/TR/performance-timeline/#observe-method, step 7.5.3.

This change exposed some inconsistencies in the logic that sends navigation
entries to PerformanceObservers, so a minor rework was required. In the new
implementation, the navigation entry is only exposed to PerformanceObservers
after the load event is handled. This is done in
LocalDOMWindow::dispatchLoadEvent(). The previous implementation did this
in FrameLoader::checkLoadCompleteForThisFrame(), with less consistent behavior.
Notably, web-platform-tests/navigation-timing/po-navigation.html seems to
fail randomly if opened outside of testing.

The spec is unclear with regards to when a navigation performance entry
should be considered available. In this implementation, there is some span
of time before the onload event in which the navigation entry is available
through performance.getEntries() while not being observable through
PerformanceObservers. This inconsistency is reasonable: the navigation entry
contains timing information about the onload event itself, so can only be
considered complete after that. Considering PerformanceObservers were
introduced to reduce unnecessary processing, it makes sense for them to
only receive completed entries.

New tests added:
* LayoutTests/imported/w3c/web-platform-tests/performance-timeline/buffered-does-not-sync-invoke.html :
    verifies that calling observe() does not invoke the callbacks synchronously. This is a direct test
    of the bug that motivated the patch.
* LayoutTests/performance-api/performance-observer-no-duplicate-navigation.html : exercises a latent
    bug (now removed) in which navigation entries were queued twice

Canonical link: https://commits.webkit.org/299508@main