[Liquid Glass] [iOS] reddit.com: no blur at bottom of "Unreviewed content" overlay https://bugs.webkit.org/show_bug.cgi?id=298324 rdar://158543289
Reviewed by Simon Fraser.

For the purposes of color sampling, the heuristic currently "punches through" transparent (or
nearly-transparent) containers with `pointer-events: none;`, that don't have any children or
background. However, this fails to account for the case where such a container has a backdrop filter
(and contains no other rendered content or background). On Reddit, this causes the "Unreviewed
content" modal overlay to be ignored for the purposes of extending fixed colors.

Handle this scenario by considering a container with CSS `backdrop-filter` to be visible, therefore
avoiding the `isHiddenOrNearlyTransparent` check. This results in extending the page background
color into the obscured edges beyond the viewport.

* LayoutTests/fast/page-color-sampling/color-sampling-full-viewport-backdrop-filter-expected.txt: Added.
* LayoutTests/fast/page-color-sampling/color-sampling-full-viewport-backdrop-filter.html: Added.
* Source/WebCore/page/LocalFrameView.cpp:
(WebCore::isHiddenOrNearlyTransparent):

Canonical link: https://commits.webkit.org/299509@main