[JSC] WASM IPInt SIMD: use the standard entry point when --useWasmIPIntSIMD https://bugs.webkit.org/show_bug.cgi?id=298363 rdar://159808955
Reviewed by Mark Lam.

Until (and unless) we have a good reason for a separate entrypoint
for functions that use SIMD, let's use the standard entrypoint
when --useWasmIPIntSIMD is enabled.

Later I'll probably rename the "SIMD" entrypoint (unless we delete
it soon) since that's now confusingly named since that's the
entrypoint that tiers up to BBQ rather immediately to bypass SIMD
execution in the interpreter. Add some asserts to clarify the usage
of the related code.

Canonical link: https://commits.webkit.org/299559@main