[JSC] Stop looking into replacement in entry tierup in BBQ https://bugs.webkit.org/show_bug.cgi?id=298368 rdar://159820765
Reviewed by Justin Michaud.

As the same to IPInt, BBQ entry point should not look for the compiled
OMG callee since we are not coming BBQ if it exists. There can be a
race, but we do not need to have precise information for using OMG here.

* Source/JavaScriptCore/wasm/WasmOperations.cpp:
(JSC::Wasm::triggerOMGReplacementCompile):
(JSC::Wasm::JSC_DEFINE_NOEXCEPT_JIT_OPERATION):

Canonical link: https://commits.webkit.org/299561@main