[ macOS Release ] imported/w3c/web-platform-tests/html/dom/idlharness.worker.html is a flakey text failure https://bugs.webkit.org/show_bug.cgi?id=296944 rdar://157586713
Reviewed by Per Arne Vollan and Youenn Fablet.

The test is flaky due to an intermittent "Blocked access to external URL" message which gets
logged by WebKitTestRunner when trying to access the internet in layout tests. Ideally, we
would be able to silence this message with "DumpJSConsoleLogInStdErr" like we usually do.
However, this is message is not generated with `console.log`, it was not getting silenced.

In this PR, I update WKTR to issue the "Blocked access to external URL" messages using
`console.log` so that they properly get silenced via "DumpJSConsoleLogInStdErr". This allows
us to easily deal with flakiness caused by such messages.

Note that imported/w3c/web-platform-tests/html/dom/idlharness.worker.html was already marked
as "DumpJSConsoleLogInStdErr", which is why I did not have to do it in this PR.

* LayoutTests/*: Rebaseline.
* Tools/DumpRenderTree/mac/ResourceLoadDelegate.mm:
(-[ResourceLoadDelegate webView:resource:willSendRequest:redirectResponse:fromDataSource:]):
* Tools/WebKitTestRunner/InjectedBundle/InjectedBundlePage.cpp:
(WTR::InjectedBundlePage::willSendRequestForFrame):

Canonical link: https://commits.webkit.org/299586@main