REGRESSION(298691@main): [ macOS wk2, iOS ] imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-back-bfcache.html https://bugs.webkit.org/show_bug.cgi?id=297477 rdar://158434206
Reviewed by Chris Dumez.

This fix the flakiness of this test.

WPT test utility has a bug that can fixed in this patch. The fetch is used for the communication
between the pages, fetch doesn't have 'keepalive' flag on, so they sometimes get cancelled because
of the navigation.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/common/dispatcher/dispatcher.js:
(async await):
(const.sendItem):
* LayoutTests/imported/w3c/web-platform-tests/common/dispatcher/remote-executor.html:

Canonical link: https://commits.webkit.org/299594@main