http/tests/security/service-worker-network-load.html is randomly timing out since the beginning rdar://116882425 https://bugs.webkit.org/show_bug.cgi?id=262734
Reviewed by Brady Eidson and Sihui Liu.

The results on the bots show that the preloadResponse promise gets resolved with undefined.
This should only happen when navigation preload is not enabled.
The script is not waiting for the worker to be activated, while the worker is enabling navigation preload within the activate event handler.
This triggers a race as the routine wait_state will sometimes resolve with 'activating' instead of 'activated'.
To remove this race, we fix the routine wait_state to actually wait for the 'activated' state.

* LayoutTests/TestExpectations:
* LayoutTests/http/tests/security/service-worker-network-load.html:

Canonical link: https://commits.webkit.org/299599@main