REGRESSION (291055@main): Buttons with box shadow are broken https://bugs.webkit.org/show_bug.cgi?id=298341 rdar://159888287
Reviewed by Aditya Keerthi.

The bug occurred with inset shadows that have large spread. The check that
the inner "hole" was empty was failing, because it checked the outer, instead
of the inner shape on the hole BorderShape.

This regressed with the adoption of BorderShape here (291055@main).

* LayoutTests/fast/box-shadow/inset-shadow-large-spread-expected.html: Added.
* LayoutTests/fast/box-shadow/inset-shadow-large-spread.html: Added.
* Source/WebCore/rendering/BackgroundPainter.cpp:
(WebCore::BackgroundPainter::paintBoxShadow const):

Canonical link: https://commits.webkit.org/299603@main