workers/empty-post-message-service-workers-crash.html spins hundreds of workers and causes flakiness https://bugs.webkit.org/show_bug.cgi?id=298424
Reviewed by Rupin Mittal.

The test spins 50 dedicated workers, each of these workers would recursively
relaunch a worker with the same script so it would infinitely launch new
scripts. This very large number of workers created could lead to flakiness
such as rdar://158179134.

To address the issue, the test now creates 5 workers, each of them now only
launches a single worker. We should thus end up with 10 workers (and 10 worker
threads) which is a lot more reasonable in terms of workload.

I am also dropping the audio element creation since it is not related to what
the test is covering and it was significantly slowing down the test as well.

* LayoutTests/workers/empty-post-message-service-workers-crash.html:
* LayoutTests/workers/empty-post-message-service-workers-crash.js:
* LayoutTests/workers/empty-post-message-service-workers-crash2.js: Added.
(onmessage):

Canonical link: https://commits.webkit.org/299616@main