[GTK][WPE] Gamepad press buttons or axis events not triggered https://bugs.webkit.org/show_bug.cgi?id=298310
Reviewed by Carlos Garcia Campos.

When WebKitGTK or WPEWebKit try to map a button or axis number to the
canonical layout defined in the W3C Gamepad spec, they return 'Unknown' in
case the button or axis number doesn't correspond to a standard value.
As a result of that, press button or axis events are not handled if
value was 'Unknown', what gives the impression the event didn't even happen.

The W3C Gamepad spec recommends to remap buttons and axis to a canonical
ordering if possible. In case not, the devices should still be exposed in
their raw form.

So, instead of returning 'Unknown' for a value outside of range,
we return the modulo of the raw value by the total number of standard
values.  That ensures the value will always be within the range and the
events will always be handled if they were triggered.

* Source/WebCore/platform/gamepad/manette/ManetteGamepad.cpp:
(WebCore::toStandardGamepadAxis):
(WebCore::toStandardGamepadButton):
(WebCore::ManetteGamepad::ManetteGamepad):
(WebCore::ManetteGamepad::buttonPressedOrReleased):
(WebCore::ManetteGamepad::absoluteAxisChanged):
* Source/WebCore/platform/gamepad/manette/ManetteGamepad.h:
* Source/WebKit/WPEPlatform/wpe/WPEGamepadManette.cpp:
(wpeGamepadButton):
(wpeGamepadAxis):
(wpeGamepadManetteStartInputMonitor):
(wpeGamepadButtonFromEvent): Deleted.

Canonical link: https://commits.webkit.org/299618@main