[macOS] -webkit-user-select: none disables searching in Safari. https://bugs.webkit.org/show_bug.cgi?id=298345 rdar://8081660
Reviewed by Aditya Keerthi.

Follow up work for mac in relation to 298785@main.

We do not have the fallback range available in the same
way on macOS as we did on iOS. In order to have a similar
solution for creating an indicator based on the range,
we need to pass that range back to the FindController so
that it can be used in the fallback case.

We also need to be checking the normalized range,
as the regular range will return the uncollapsed version,
and that will not send us down the fallback path.

* Source/WebCore/editing/Editor.cpp:
(WebCore::Editor::findString):
* Source/WebCore/editing/Editor.h:
* Source/WebCore/editing/EditorCommand.cpp:
(WebCore::executeFindString):
* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::find const):
* Source/WebCore/page/Page.cpp:
(WebCore::Page::findString):
* Source/WebCore/page/Page.h:
* Source/WebKit/WebProcess/WebPage/FindController.cpp:
(WebKit::FindController::frameWithSelection):
(WebKit::FindController::findString):
(WebKit::FindController::updateFindIndicator):
* Source/WebKit/WebProcess/WebPage/FindController.h:

Canonical link: https://commits.webkit.org/299624@main