[JSC] Skip more test in 32-bit JSC. https://bugs.webkit.org/show_bug.cgi?id=298456
Reviewed by Sosuke Suzuki.

For wasm-js-string-builtins.js we need to properly handle argument registers,
as the implementation always assumes 1 register per argument, which is not true
for 32-bit JSC. We need to change the implementation to also spill arguments,
which is currently not implemented in either IPInt or JITed code gen.

The other 3 need IPInt to work properly.

* JSTests/wasm/lowExecutableMemory/executable-memory-oom.js:
* JSTests/wasm/stress/cc-int-to-int-cross-module-with-exception.js:
* JSTests/wasm/stress/cc-int-to-int-jit-to-llint.js:
* JSTests/wasm/stress/wasm-js-string-builtins.js:

Canonical link: https://commits.webkit.org/299634@main