KeyError: '8e2703a62c3ac7282b9f96744a53646a89b59c9a' when running 'git-webkit pr' https://bugs.webkit.org/show_bug.cgi?id=275693
Reviewed by Jonathan Bedard.

Check if the hash exists in the self._hash_to_identifiers table before
attempting to use it, to avoid a KeyError exception if it's not present.
Looks like we're generally pretty careful to check whether the value
exits in the table before using it, but are missing a check here.

Canonical link: https://commits.webkit.org/299639@main