[field-sizing] Larger placeholder font-size should expand the field https://bugs.webkit.org/show_bug.cgi?id=269127
Reviewed by Antti Koivisto.

RenderTextControl::computeLogicalHeight is supposed to return a hugging size when field-sizing is on.
It essentially means stretching the height to the size of its content.
Normally (bfc) computeLogicalHeight happens on the "way back", meaning that the renderer's subtree is laid out by then
and we could just consult the placeholder's content height.

While placeholder is part of the text control's subtree, it is an excluded renderer (not taking in-flow space, positioned by hand)
and we don't lay it out until after we start positioning the inner renderer(s) for the input _after_ layout (still in RenderTextControlSingleLine::layout though)

This fix initiates layout on this excluded placeholder in ::layoutExcludedChildren (duh), and includes its height when computing the input box's final height.

* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-number-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-text-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-placeholder-stretch-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-placeholder-stretch-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-placeholder-stretch.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-textarea-expected.txt:
* LayoutTests/platform/ios-18/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-textarea-expected.txt:
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-number-expected.txt:
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-text-expected.txt:
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-textarea-expected.txt:
* LayoutTests/platform/mac-sequoia-wk2/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-number-expected.txt:
* LayoutTests/platform/mac-wk2/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-number-expected.txt:
* Source/WebCore/rendering/RenderTextControl.cpp:
(WebCore::RenderTextControl::computeLogicalHeight const):
(WebCore::RenderTextControl::layoutExcludedChildren):

Canonical link: https://commits.webkit.org/299672@main