Reduce size of Development sandboxes https://bugs.webkit.org/show_bug.cgi?id=298469 rdar://158907488
Reviewed by Sihui Liu.

Currently, the development sandboxes include their original counterpart, and replace blocking rules with
rules that allow the operation, but emit a report. This causes the development sandboxes to be larger
than necessary, since the original sandboxes have complexity that is not required in the development
versions. This patch creates a custom development sandbox for each process, and only allows the
operations in developer mode.

* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.Development.sb.in:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.GPU.sb.in:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.Networking.Development.sb.in:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.Networking.sb.in:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.Development.sb.in:
* Source/WebKit/Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb.in:
* Source/WebKit/Shared/Sandbox/iOS/common.sb:
* Source/WebKit/Shared/Sandbox/iOS/gpu-defines.sb:
* Source/WebKit/Shared/Sandbox/iOS/networking-defines.sb:
* Source/WebKit/Shared/Sandbox/iOS/webcontent-defines.sb:

Canonical link: https://commits.webkit.org/299698@main