AX: For dynamic changes requiring a full-object update, swap out IsolatedObjectData rather than fully destroying and recreating a new AXIsolatedObject to avoid stale object bugs https://bugs.webkit.org/show_bug.cgi?id=298481 rdar://159975526
Reviewed by Joshua Hoffman.

For some types of dynamic page updates, like value changes in text fields or accessibility text changes, we will not
try to invalidate individual properties and instead create a full-object update. We do this by creating a new
AXIsolatedObject, and swapping out the wrapper and AXIsolatedTree::m_readerThreadNodeMap membership in AXIsolatedTree::applyPendingChanges.

The downside of this approach is that just because an AXIsolatedObject is removed from AXIsolatedTree::m_readerThreadNodeMap
doesn't mean it will be destroyed, as various persistent data structures can hold a strong reference to the old AXIsolatedObject.
One of the most prevalent examples of this is an object's m_children. This means it's possible to use the old, stale version
of the object to serve accessibility requests, which can result in incorrect behavior.

Fix this by only swapping out IsolatedObjectData for an object in these full-update circumstances, rather than creating
a whole new AXIsolatedObject.

* LayoutTests/accessibility/dynamic-selected-radio-button-expected.txt: Added.
* LayoutTests/accessibility/dynamic-selected-radio-button.html: Added.
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::updateFromData):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.h:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::applyPendingChangesLocked):

Canonical link: https://commits.webkit.org/299712@main