REGRESSION(299701@main): Broke watchOS and tvOS builds https://bugs.webkit.org/show_bug.cgi?id=298553 rdar://160143108
Unreviewed build fix.

watchOS and tvOS do not support pointing devices, and as such the
UITouch.rollAngle API is not exposed on said platforms. In 299701@main,
we were reading the roll angle unconditionally (on iOS), which lead to
build failures like this:

```
'rollAngle' is unavailable: not available on watchOS
WKTouchEventsGestureRecognizer.mm:352
'rollAngle' is unavailable: not available on watchOS
WKTouchEventsGestureRecognizer.mm:266
```

We fix the build failure in this patch by gating this behavior to
platforms that do ship with rollAngle support, falling back to the
default twist value when appropriate.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WTF/wtf/PlatformHave.h:
* Source/WebKit/UIProcess/ios/WKTouchEventsGestureRecognizer.mm:
(rollAngleOrDefault):
(-[WKTouchEventsGestureRecognizer _touchEventForChildTouch:withParent:]):
(-[WKTouchEventsGestureRecognizer _recordTouches:ofType:forEvent:]):

Canonical link: https://commits.webkit.org/299713@main